refactor(chapter06): rename files
